requirements.txt
flask
python-dotenv
groq
fpdf